Added rewrite functionality to the Server
Rewrite rules can now be specified in the server that will transform the
URL prior to it being passed to the router.

This occurs before the router is invoked or middleware is processed.

Rewrite was implemented on the actual server, as apposed to middleware,
due the way that scopes are processed. The scope's middleware is not
invoked unless the URL has already been verified to exist within that
scope.

Resolves: #9