Method based routing (#13)
Routing should obey the http method in the request. Previously the
method was ignored and route matching was done strictly on path. This
has been changed so that a route no longer matches if the http method
verb is not correct. This will cause URLs hit with a method different
from what is specified in the routes to return a 404 error.

This changed forced a change to the scope and router API. The handle
functions now accept the http.Request instead of a path string.
Middleware and the testing package were updated to compensate.

Resolves #10