Request IDs
Added UUID generation for unique requestIDs. These IDs are automatically
generated when a context is created and appended to the request output.
This is intended to be used for log tracing.

The server tests were also refactored, as they were matching the entire
response text against static text. This wont work since the requestIDs
are now randomly generated.

All the tests have been changed to use struct level validation.