PreRoute middleware
Scopes now have a concept of "Pre" middleware that evaluates before a
scopes routes are matched.

This allows middleware to make changes to the context before the routes
are checked or evaluated. Pre middleware happens before the route
checking occurs, which means these middleware will fire even for
routes that don't exist.

This was done specifically for things like CORS and Rewrite middleware.
Where the destination route doesn't exist but middleware should still be
ran.