Added basic static file serving
Scopes now have a ServerPath function that can be used to automatically
generate a route that returns files from a path.

A File function was added to the context to return a file response. This
is handled in ServeHTTP if a file is set on the response the file is
served directly instead of the normal processing by the ResponseAdapter