Universal Logging (#51)
Added a log instance to the server which is passed to the Context. This gives endpoints the access to a unified rolling application log.

Refactored the way the request logger middleware works. It now uses an internal Vox instance to output logs which is provided inside the configuration structure.

This should allow more control of the output stream now that it can take advantage of vox output pipelines.

Resolves #45