Basic Channel (websocket) Implementation (#56)
A fully featured web socket implementation that is internally built into the router and handlers.

A channel helper provides a socket based route with a specialized route handler to handle web socket communication. Various events are passed to this handler including Connect, Disconnect, and Message events.

Each channel maintains a list of its clients and messages can be sent directly or broadcast.

A room system has also been implemented which allows arbitrary groups to be created within a channel and clients can be added and removed from them. Rooms also maintain their own client lists and have broadcast functionality.

As part of the channel implementation the server reference is now passed into the scopes and down to the context. This allows normal endpoints to reach out to the server and get a hold of the list of channels to send messages.